# api/gen/go/server/service/v1/i_model_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: server/service/v1/i_model.proto

package serverpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-admin/api/gen/go/model/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationModelServiceCreate = "/server.service.v1.ModelService/Create"
const OperationModelServiceDelete = "/server.service.v1.ModelService/Delete"
const OperationModelServiceExists = "/server.service.v1.ModelService/Exists"
const OperationModelServiceGet = "/server.service.v1.ModelService/Get"
const OperationModelServiceList = "/server.service.v1.ModelService/List"
const OperationModelServiceListObjects = "/server.service.v1.ModelService/ListObjects"
const OperationModelServiceListTags = "/server.service.v1.ModelService/ListTags"
const OperationModelServiceUpdate = "/server.service.v1.ModelService/Update"

type ModelServiceHTTPServer interface {
	// Create 创建模型
	Create(context.Context, *v11.CreateModelRequest) (*v11.Model, error)
	// Delete 删除模型
	Delete(context.Context, *v11.DeleteModelRequest) (*emptypb.Empty, error)
	// Exists 是否存在
	Exists(context.Context, *v11.ModelExistsRequest) (*v11.ModelExistsResponse, error)
	// Get 模型详情
	Get(context.Context, *v11.GetModelRequest) (*v11.Model, error)
	// List 模型列表
	List(context.Context, *v1.PagingRequest) (*v11.ListModelResponse, error)
	// ListObjects 模型对象列表
	ListObjects(context.Context, *v11.ListObjectsRequest) (*v11.ListObjectsResponse, error)
	ListTags(context.Context, *v11.ListTagsRequest) (*v11.ListTagsResponse, error)
	// Update 更新模型
	Update(context.Context, *v11.UpdateModelRequest) (*emptypb.Empty, error)
}

func RegisterModelServiceHTTPServer(s *http.Server, srv ModelServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/server/v1/models", _ModelService_List1_HTTP_Handler(srv))
	r.GET("/server/v1/models/{id}", _ModelService_Get1_HTTP_Handler(srv))
	r.POST("/server/v1/models", _ModelService_Create1_HTTP_Handler(srv))
	r.PUT("/server/v1/models/{id}", _ModelService_Update1_HTTP_Handler(srv))
	r.DELETE("/server/v1/models/{id}", _ModelService_Delete1_HTTP_Handler(srv))
	r.GET("/server/v1/models:exists", _ModelService_Exists0_HTTP_Handler(srv))
	r.GET("/server/v1/models/{id}/tags", _ModelService_ListTags0_HTTP_Handler(srv))
	r.GET("/server/v1/models/{id}/repositories/{branch}/objects/ls", _ModelService_ListObjects0_HTTP_Handler(srv))
}

func _ModelService_List1_HTTP_Handler(srv ModelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListModelResponse)
		return ctx.Result(200, reply)
	}
}

func _ModelService_Get1_HTTP_Handler(srv ModelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetModelRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetModelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Model)
		return ctx.Result(200, reply)
	}
}

func _ModelService_Create1_HTTP_Handler(srv ModelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateModelRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreateModelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Model)
		return ctx.Result(200, reply)
	}
}

func _ModelService_Update1_HTTP_Handler(srv ModelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateModelRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateModelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ModelService_Delete1_HTTP_Handler(srv ModelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteModelRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteModelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _ModelService_Exists0_HTTP_Handler(srv ModelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.ModelExistsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelServiceExists)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Exists(ctx, req.(*v11.ModelExistsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ModelExistsResponse)
		return ctx.Result(200, reply)
	}
}

func _ModelService_ListTags0_HTTP_Handler(srv ModelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.ListTagsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelServiceListTags)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListTags(ctx, req.(*v11.ListTagsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListTagsResponse)
		return ctx.Result(200, reply)
	}
}

func _ModelService_ListObjects0_HTTP_Handler(srv ModelServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.ListObjectsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelServiceListObjects)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListObjects(ctx, req.(*v11.ListObjectsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListObjectsResponse)
		return ctx.Result(200, reply)
	}
}

type ModelServiceHTTPClient interface {
	// Create 创建模型
	Create(ctx context.Context, req *v11.CreateModelRequest, opts ...http.CallOption) (rsp *v11.Model, err error)
	// Delete 删除模型
	Delete(ctx context.Context, req *v11.DeleteModelRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Exists 是否存在
	Exists(ctx context.Context, req *v11.ModelExistsRequest, opts ...http.CallOption) (rsp *v11.ModelExistsResponse, err error)
	// Get 模型详情
	Get(ctx context.Context, req *v11.GetModelRequest, opts ...http.CallOption) (rsp *v11.Model, err error)
	// List 模型列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListModelResponse, err error)
	// ListObjects 模型对象列表
	ListObjects(ctx context.Context, req *v11.ListObjectsRequest, opts ...http.CallOption) (rsp *v11.ListObjectsResponse, err error)
	ListTags(ctx context.Context, req *v11.ListTagsRequest, opts ...http.CallOption) (rsp *v11.ListTagsResponse, err error)
	// Update 更新模型
	Update(ctx context.Context, req *v11.UpdateModelRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type ModelServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewModelServiceHTTPClient(client *http.Client) ModelServiceHTTPClient {
	return &ModelServiceHTTPClientImpl{client}
}

// Create 创建模型
func (c *ModelServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreateModelRequest, opts ...http.CallOption) (*v11.Model, error) {
	var out v11.Model
	pattern := "/server/v1/models"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationModelServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Delete 删除模型
func (c *ModelServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteModelRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/server/v1/models/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationModelServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Exists 是否存在
func (c *ModelServiceHTTPClientImpl) Exists(ctx context.Context, in *v11.ModelExistsRequest, opts ...http.CallOption) (*v11.ModelExistsResponse, error) {
	var out v11.ModelExistsResponse
	pattern := "/server/v1/models:exists"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationModelServiceExists))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Get 模型详情
func (c *ModelServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetModelRequest, opts ...http.CallOption) (*v11.Model, error) {
	var out v11.Model
	pattern := "/server/v1/models/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationModelServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 模型列表
func (c *ModelServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListModelResponse, error) {
	var out v11.ListModelResponse
	pattern := "/server/v1/models"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationModelServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListObjects 模型对象列表
func (c *ModelServiceHTTPClientImpl) ListObjects(ctx context.Context, in *v11.ListObjectsRequest, opts ...http.CallOption) (*v11.ListObjectsResponse, error) {
	var out v11.ListObjectsResponse
	pattern := "/server/v1/models/{id}/repositories/{branch}/objects/ls"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationModelServiceListObjects))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ModelServiceHTTPClientImpl) ListTags(ctx context.Context, in *v11.ListTagsRequest, opts ...http.CallOption) (*v11.ListTagsResponse, error) {
	var out v11.ListTagsResponse
	pattern := "/server/v1/models/{id}/tags"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationModelServiceListTags))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新模型
func (c *ModelServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateModelRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/server/v1/models/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationModelServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}